feat(FEC-7814): add setAdWillAutoPlay and setAdWillPlayMuted support
Following the new autoplay changes in Chrome and other browsers IMA are making autoplay and muted playback as key attributes in their attribution and ad buying process.
As part of this they are requesting that the following APIs will be set accordingly so this data, if ad will be played autoplay or/and muted, will be passed to the ad server.
Setting of the data is via API:
[setAdWillAutoPlay](https://developers.google.com/interactive-media-ads/docs/sdks/html5/v3/apis#ima.AdsRequest.setAdWillAutoPlay)
[setAdWillPlayMuted](https://developers.google.com/interactive-media-ads/docs/sdks/html5/v3/apis#ima.AdsRequest.setAdWillPlayMuted)
See:
https://github.com/googleads/googleads-ima-html5/blob/master/attempt_to_autoplay/ads.js